Running tickerv2 for the moment. more details to look at, more modification needed as per to peers updates and changes
